font-patcher: Really set SubFamily to what we want
[why]
The most convenient way to set something in the SFNT table is by using
Fontforge's appendSFNTName(). When we try to set the SubFamily, this
will succeed in many instances, but not in all. Sometimes the function
seems to be ignored.

This is for example the case for VictorMono-Medium.ttf where we want to
set the SubFamily "Regular", but the final patched file has the
SubFamily "Medium".

[how]
The reason for the unexpected behavior of Fontforge is as follows. It
almost seems like a bug, but then it is coded explicitely in this way.
Maybe noone ever stumbled about it, or it has only been designed only
with creating new fonts in mind.

Fontforge lets you set any value, unless it is the default value. If it
is the default value it does not set anything. It also does not remove
a previously existing non-default value. Why it is done this way is
unclear:
  fontforge/python.c SetSFNTName() line 11431
    return( 1 ); /* If they set it to the default, there's nothing to do */

Then is the question: What is the default? It is taken from the
currently set fontname (??!). The fontname is parsed and everything
behind the dash is the default SubFamily:
  fontforge/tottf.c DefaultTTFEnglishNames()
  fontforge/splinefont.c _GetModifiers()

To stick to the example above, the fontname is there
"VictorMonoNerdFontM-Medium". Afterwards we want to set the SubFamily to
"Medium" - which is the 'default' and Fontforge decides that nothing has
to be done. Albeigth the current SubFamily is "Regular" (from the source
font load).

Well.

To fix this without touching Fontforge we need to set the SubFamily
directly in the SFNT table, which is a little bit unhandy but possible.

Now in the generated files the SubFamily is again as expected the same
es the fontname's last part.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>