name_parser: Work around fontforge limitations
[why]
The fontforge Python interface is not able to easily remove existing
SFNT entries.

[how]
Because it it not so easy to apply the findings of the FontnameParser to
a font a helper function is added that does all the necessary bits and
pieces.

[note]
Also fix instances where name_parser uses Python features that are
too 'new' on some machines (i.e. Gitlab Ubu 18.04 runner).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>