name-parser: Strip suffixes
[why]
Suffixes should not start with a blank or end with one, the resulting
names would look wrong. But `font-patcher` calculates them with a
beginning blank (for easy addition in the old code).

It is more convenient and also more robust if the parser makes sure that
no superfluous blanks are added.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>