font-patcher: Use FontnameParser optionally
[why]
To correct a lot of problems we want to use a new way to set the
various name parts in the patched font.

[how]
Keep the old behavior, add new command line flag that enables the new
method (i.e. --parser).

The result will be unchanged if the flag is not given.

The FontnameParser is only used to set the names if the flag is given.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>