name_parser: Add option to create (marginally) more concise Family names
[why]
The font-patcher historically grouped all name parts of the original
font together (if there are any) to form a camel case name:
  Cascadia Code => CaskaydiaCove

But only in the Family and Preferred Family names.

[how]
We already have an option to create short(er) Families by usage of
abbreviated style names (stemming from Noto).

Add an additional option that CamelCases the basic fontname in the
Family names.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>