name_parser: Add script to show differences to old behavior
The script patches a given font two times with font-patcher
* 'normal', just with --powerline and noithing more
* one times with --parser specified (will use the FontnameParser

The font files that resulted will be opened and the embedded names
compared. It is important to really patch a file and see the outcome
because fontforge is a bit unpredictable when it comes to what actuall
is set in a font file ;)

The names are compared with a but of lenience, like with
name_parser_test1.

What is missing is a known-issues file. Will add that later.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>