font info: Remove references to 'Terminess' from non-Terminus fonts
[why]
In some fonts, on in the `patched-fonts/` directory the font information
files contain an explanation why the font is called "Terminess" and not
"Termius" - but the font is neither.

Affected: DroidSansMono and Ubuntu (non mono)

I assume this is a 'I copied the info file over from an existing font
and just change the names' (but the source was unfortunately Terminess).

[how]
Manually remove the paragraph from all files that contain it which but
have another name altogether.

[note]
Not shure if they are autogenerated somehow, with for example
`standardize-and-complete-readmes.sh` ... but a `git grep` showed only
the files that are changed with this PR.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>