font-patcher: Remove CV lookups in Iosevka
[why]
For some reason fontforge is unable to pack all the lookup tables of
Iosevka back into the generated font, although they fit obviously in the
source font.

The lookup tables will become garbled and fontforge says something like:

  Attempt to output 67666 into a 16-bit field.
  It will be truncated and the file may not be useful.

As this also happens if we do nothing, just open and generate the font
without touching, there is not much we can do.

[how]
The only thing we can do is reduce the number of lookup tables that are
in the font file. This is somewhat brute, but at least we end up with a
usable font, albeit with a little bit less features.

Unfortunately the SS (Style Set) tables are not enough, but the CV
(Character Variants) suffice. So all CV lookups are riped out.

Fixes: #694

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>