font-patcher: Increase h-overlap, decrease v-overlap
[why]
On some machine the gap is still visible, but with 4% overlap the
result looks good.

[how]
Increase the overlap from 3% to 4%.

Cascadia Code has 8% of overlap, but they have a special extension to
the overlap side, we will have a visual glitch if the overlap becomes
too big. So lets stay with 4% or we should change the basis for the
powerline glyphs.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>