previews: Update Iosevka image
[why]
The Iosevka image is somehow broken, the dots of `i` and `j` are
shifted. Maybe the tool to generate the image had a problem with the
partially broken Iosevka font? With current tools but the same v2.1.0
font there is no problem.

[how]
Create the image with generate-font-image-previews.sh. But I have no
`svgo` (svg optimizer), so it is unoptimized.
Also afterwards scale the 'page' down and save as plain svg in
Inkscape.

Fixes: 865

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>