Update 3270font to v3.0.1
[why]
We have only a rather old 3270font: v1.2.23

This new font can be processed with current FontForge without problems.
And we have the benefit of an update.

[how]
Because the paths change (Medium->Regular etc) we need to purge the
patched-fonts/ directories manually. The automatic purging does work
only if at least the paths are kept.

[note]
https://github.com/rbanffy/3270font

Fixes: #947

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>