font-patcher: Make (some) powerline glyphs always monospaced
[why]
When a non-Mono font (i.e. `Nerd Font` not `Nerd Font Mono`) is created
the goal is to have big (up to 2 cell wide) symbols. But usually you
still want to have only one cell wide powerline glyphs, to have them
handable.

[how]
Introduce a new attribute parameter 'single' that means that the (xy in
the changed cases) scaling to one cell shall take place even if `-s` is
not given.

It is ignored if the sourcefont is not monospaced, because we could not
scale in that case anyhow (to which cell-size should we scale?).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>