font-patcher: Remove overlap for line style powerline glyphs
[why]
The overlap into the previous/next character is used to prevent small
gaps in colored powerlines. These show up as small colored line and can
usually be suppressed by changing the antialiasing method. But for
people who can no change the antialiasing the overlap 'covers up' the
gap.

But this is not relevant if the glyph is just a 'line'. There just isn't
any gap that needs covering-up.

[note]
Also: Add missing RHS-waveform instructions.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>