download: Allow patched and unpatched name to differ with blanks
[why]
We detect if a font has a reserved font name RFN by comparing unpatched
and patched names.

In the patched names we remove blanks from the original name, example
"BigBlue Terminal" -> "BigBlueTerminal Nerd Font"

In the fonts.json database we store "BigBlueTerminal" as `.patchedName`.
For our RFN detection we need to set the `.unpatchedName` also to
"BigBlueTerminal", which is strictly speaking not correct.

[how]
Compare `.unpatchedName` and `.patchedName` without taking the blanks into
account: Remove them before comparing.

[note]
This is needed to finally have the correct original names in the
database, which will be used for the Casks creation.

[note]
Also unify the shell commands (remove whitespaces).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>