download: Fix RFN detection
[why]
Sometimes we change the font name just because ... reasons, without
having being oblieged to do that because of a reserved font name (RFN)
clause.

This is for example the case with M+ that we rename to MPlus (well, it
has both names in principle, but the human readable name in the original
fonts in M+).

Another is "Go Mono" that we rename to "Go-Mono" for whatever reason.

This would also have been needed when `Fira Code` was still RFN and we
got the special permission to patch it without dropping `Fira Code` from
the name (see [1]), to create the correct tagged list.
But Fira Code dropped the RFN some time ago, so no need for this font
anymore.

This also prevents our own Symbols Only fonts to be listed as having RFN.

[how]
But an explicit field into the fonts.json database if the font has an
RFN or not, and decide upon that.

[1] https://github.com/tonsky/FiraCode/issues/767

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>