casks: Fix naming of Symbols Only font
[why]
The symbols only font has two entries in the fonts.json and that breaks
generation of the cask.

[how]
Do the queries on a 'first wins' strategy; ignore further entries in the
fonts.json.

[note]
I guess we should remove one of the two instances of Symbols Only from
the database; I have still not figured out why there are two distinct
entries.
And why there are two fonts at all :grimacing:

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>