font-patcher: Remove overlap from line-ish glyphs
[why]
Most Powerline glyphs have a little bit over overlap to the previous or
next glyph to prevent a 'break' in a colored prompt.

It does not make sense to have overlap with glyphs that can never
produce any of that issues, i.e. glyphs that are not filled to the
border. Like all the line-ish glyphs.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>