download: Fix RFN detection for un-renamed fonts
[why]
When we do not rename a font that has the RFN flag set we do not list
the font as RFN-font.
This is for example the case with mononoki since the last commit.

[how]
Show the source font name not only when it differs but also when it is a
RFN.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>