font-patcher: Optimize add_glyphrefs_to_essential
[why]
Using `continue` feels inelegant when there's a way to write the if
conditions in add_glyphrefs_to_essential() without necessitating the use
of `continue` while ensuring that the function still works as intended.

[how]
Change the `if` conditions and remove any usage of the `continue`
keyword in add_glyphrefs_to_essential().