cheat-sheet: Allow click to copy on hex value
[why]
We removed the functionality with the previous commit and replaced it
with 'copy the UTF16 code' (needed for codepoints > u_FFFF).

But people might want to have the possibility to easily (by click) copy
the codepoint hex number.

[how]
Add highlighting on the hex number and allow clicking to copy the value.

[note]
The tooltip will be added in the cheat-sheet generator.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>