Fix: Fix JetBrains Mono Family names (repatch)
[why]
The JetBrains Mono family names are still broken.

The commit
  b98c2a19  Fix JetBrains Mono Family names

did not do much, because it was not activated with gotta-patch-em.
The previous commit fixed that.

Put the fonts in the repo are still wrong. Repatching those gives a
chance to test them properly before v3.0.0.

[how]
Run gotta-patch-em with version set to v3.3.3-p6 because thats the git
status right now.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>