SourceCodePro: Remove obsolete font
[why]
In the course of #271 the source for SourceCodePro switched: Before we
used SourceCode Powerline [1], afterwards the original one [2].

One font file has been forgotten, as the font files were differently
named in both variants.

This caused issed with parallel build in the CI and font tweaking (which
is done as additional process after the actual patching).

[note]
[1] https://github.com/ryanmcilmoyl/sourceCodePro-Powerline
[2] https://github.com/adobe-fonts/source-code-pro/releases/tag/2.030R-ro%2F1.050R-it

Maybe, this is not entirely clear, and the 'Powerline' variant does not
have an Italic style.

Fixes: #1084

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>